Improve seeding process
