+++ models/Team/Team.js
@@ -1,7 +1,7 @@
 /**
  * Example model for querying the teams
  */
-import db from '../helpers/db';
+import db from '../../helpers/db';
 
 export const getTeam = (id, done) => {
   db.get().query(`SELECT * FROM Team WHERE id=${id}`, (err, rows) => {
@@ -24,7 +24,7 @@ export const addTeam = (name, done) => {
   db.get().query(`INSERT INTO Team (name) VALUES ('${name}')`, (err, rows) => {
     if (err) return done(null, err);
     // TODO: standardize API results with success / failure, etc
-    return { result: true };
+    done(rows[0]);
   });
 };
