+++ models/Team/index.js
@@ -1,0 +1,2 @@
+import Team from './Team';
+export default Team;
