+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon index.js --exec babel-node --presets  es2015,stage-2",
-    "seed": "babel-node schema.js --presets es2015,stage-2"
+    "seed": "babel-node seed.js --presets es2015,stage-2",
+    "test": "jest"
   },
   "repository": {},
   "author": "316 Power Squad",
@@ -20,6 +21,7 @@
     "babel-cli": "^6.23.0",
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-stage-2": "^6.22.0",
+    "jest": "^21.2.0",
     "nodemon": "^1.11.0"
   }
 }
